NTV2Buffer: make sure Allocate() does a complete Fill(), force UByte
If Fill() is NOT done with a byte type there is a chance the complete buffer
will not be filled if the buffer size is not a perfect multiple of sizeof(T).

+++ ajantv2/src/ntv2publicinterface.cpp
@@ -1648,7 +1648,7 @@ bool NTV2Buffer::Allocate (const size_t inByteCount, const bool inPageAligned)
 	if (GetByteCount()	&&	fFlags & NTV2Buffer_ALLOCATED)	//	If already was Allocated
 		if (inByteCount == GetByteCount())					//	If same byte count
 		{
-			Fill(0);		//	Zero it...
+			Fill(UByte(0));		//	Zero it...
 			return true;	//	...and return true
 		}
 
@@ -1670,7 +1670,7 @@ bool NTV2Buffer::Allocate (const size_t inByteCount, const bool inPageAligned)
 			fFlags |= NTV2Buffer_ALLOCATED;
 			if (inPageAligned)
 				fFlags |= NTV2Buffer_PAGE_ALIGNED;	//	Set "page aligned" flag
-			Fill(0);	//	Zero it
+			Fill(UByte(0));	//	Zero it
 		}
 	}	//	if requested size is non-zero
 	return result;
